35 user item link

+++ back-end/src/main/java/com/ac3/InventorMe/repository/LoginRepository.java
@@ -3,6 +3,6 @@ package com.ac3.InventorMe.repository;
 import com.ac3.InventorMe.model.Login;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LoginRepository extends JpaRepository<Login, Long> {
+public interface LoginRepository extends JpaRepository<Login, Integer> {
     Login findByEmail(String email);
 }
